requirements.txt
black>=25.11.0
django>=5.2.8
django-debug-toolbar>=6.1.0
django-filter>=25.2
django-rest-framework>=0.1.0
django-stubs>=5.2.7
djangorestframework-simplejwt>=5.5.1
drf-spectacular>=0.29.0
flake8>=7.3.0
isort>=7.0.0
pillow>=12.0.0
psycopg>=3.2.13
python-dotenv~=1.2.1
djangorestframework~=3.16.1
psycopg-binary==3.2.13